+++ package.json
@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "axios": "^1.4.0",
+    "commitizen": "^4.3.0",
     "jsdoc": "^4.0.2",
+    "json-server": "^0.17.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-query": "^3.39.3",
-    "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1",
     "react-scroll": "^1.8.9",
-    "redux": "^4.2.1",
     "styled-components": "^6.0.4",
     "web-vitals": "^2.1.0"
   },
