--- src/1. mainpage/components/footer.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Footer = () => {
-  return (
-    <div>footer</div>
-  )
-}
-
-export default Footer
